Support multiple JUnit tests in BazelTestRunner
This is a PoC, rather than a fully-finished suggestion on how this
feature should be implemented!

There is currently a restriction on there being a single class as the
entry point for a set of JUnit 4 tests. This can be problematic on large
code-bases as it forces a trade-off between using multiple test workers
(which incur start-up and tear-down costs) or running all tests as a
JUnit `@Suite` (which is cumbersome and doesn't play nicely with Bazel's
built-in flaky detection.